Add files via upload
This was an earlier prototype of a finite state machine, I was trying to make my own IR protocol to increase transmission speed but it was unreliable so I used the IRremote Arduino library. I included it here as an example of what we can do, it just needs to be incorporated or modified.